Extract helpers
A first step to find common code and simplify tests.
