+++ tests/dv_backup_test.go
@@ -21,6 +21,7 @@ var kvClient *kubecli.KubevirtClient
 
 var _ = Describe("DV Backup", func() {
 	var client, _ = util.GetK8sClient()
+	var kvClient kubecli.KubevirtClient
 	var namespace *v1.Namespace
 	var timeout context.Context
 	var cancelFunc context.CancelFunc
@@ -30,10 +31,9 @@ var _ = Describe("DV Backup", func() {
 	var r = framework.NewKubernetesReporter()
 
 	BeforeSuite(func() {
-		var err error
-
-		kvClient, err = util.GetKubeVirtclient()
+		kvClientRef, err := util.GetKubeVirtclient()
 		Expect(err).ToNot(HaveOccurred())
+		kvClient = *kvClientRef
 	})
 
 	BeforeEach(func() {
@@ -54,7 +54,7 @@ var _ = Describe("DV Backup", func() {
 		}
 
 		// Deleting the backup also deletes all restores, volume snapshots etc.
-		err := DeleteBackup(timeout, backupName, r.BackupNamespace)
+		err := framework.DeleteBackup(timeout, backupName, r.BackupNamespace)
 		if err != nil {
 			fmt.Fprintf(GinkgoWriter, "Err: %s\n", err)
 		}
@@ -72,57 +72,57 @@ var _ = Describe("DV Backup", func() {
 
 		BeforeEach(func() {
 			var err error
-			dvSpec := NewDataVolumeForBlankRawImage("test-dv", "100Mi", r.StorageClass)
+			dvSpec := framework.NewDataVolumeForBlankRawImage("test-dv", "100Mi", r.StorageClass)
 			dvSpec.Annotations[forceBindAnnotation] = "true"
 
 			By(fmt.Sprintf("Creating DataVolume %s", dvSpec.Name))
-			dv, err = CreateDataVolumeFromDefinition(*kvClient, namespace.Name, dvSpec)
+			dv, err = framework.CreateDataVolumeFromDefinition(kvClient, namespace.Name, dvSpec)
 			Expect(err).ToNot(HaveOccurred())
 
-			err = WaitForDataVolumePhase(*kvClient, namespace.Name, cdiv1.Succeeded, "test-dv")
+			err = framework.WaitForDataVolumePhase(kvClient, namespace.Name, cdiv1.Succeeded, "test-dv")
 			Expect(err).ToNot(HaveOccurred())
 		})
 
 		AfterEach(func() {
-			err := DeleteDataVolume(*kvClient, namespace.Name, dv.Name)
+			err := framework.DeleteDataVolume(kvClient, namespace.Name, dv.Name)
 			Expect(err).ToNot(HaveOccurred())
 		})
 
 		It("Backup should succeed", func() {
-			err := CreateBackupForNamespace(timeout, backupName, namespace.Name, snapshotLocation, r.BackupNamespace, true)
+			err := framework.CreateBackupForNamespace(timeout, backupName, namespace.Name, snapshotLocation, r.BackupNamespace, true)
 			Expect(err).ToNot(HaveOccurred())
 
-			phase, err := GetBackupPhase(timeout, backupName, r.BackupNamespace)
+			phase, err := framework.GetBackupPhase(timeout, backupName, r.BackupNamespace)
 			Expect(err).ToNot(HaveOccurred())
 			Expect(phase).To(Equal(velerov1api.BackupPhaseCompleted))
 		})
 
 		It("DataVolume should be restored", func() {
 			By("Crating backup test-backup")
-			err := CreateBackupForNamespace(timeout, backupName, namespace.Name, snapshotLocation, r.BackupNamespace, true)
+			err := framework.CreateBackupForNamespace(timeout, backupName, namespace.Name, snapshotLocation, r.BackupNamespace, true)
 			Expect(err).ToNot(HaveOccurred())
 
-			phase, err := GetBackupPhase(timeout, backupName, r.BackupNamespace)
+			phase, err := framework.GetBackupPhase(timeout, backupName, r.BackupNamespace)
 			Expect(err).ToNot(HaveOccurred())
 			Expect(phase).To(Equal(velerov1api.BackupPhaseCompleted))
 
 			By("Deleting DataVolume")
-			err = DeleteDataVolume(*kvClient, namespace.Name, dv.Name)
+			err = framework.DeleteDataVolume(kvClient, namespace.Name, dv.Name)
 			Expect(err).ToNot(HaveOccurred())
 
-			ok, err := WaitDataVolumeDeleted(*kvClient, namespace.Name, dv.Name)
+			ok, err := framework.WaitDataVolumeDeleted(kvClient, namespace.Name, dv.Name)
 			Expect(err).ToNot(HaveOccurred())
 			Expect(ok).To(BeTrue())
 
 			By("Creating restore test-restore")
-			err = CreateRestoreForBackup(timeout, backupName, restoreName, r.BackupNamespace, true)
+			err = framework.CreateRestoreForBackup(timeout, backupName, restoreName, r.BackupNamespace, true)
 			Expect(err).ToNot(HaveOccurred())
 
-			err = WaitForRestorePhase(timeout, restoreName, r.BackupNamespace, velerov1api.RestorePhaseCompleted)
+			err = framework.WaitForRestorePhase(timeout, restoreName, r.BackupNamespace, velerov1api.RestorePhaseCompleted)
 			Expect(err).ToNot(HaveOccurred())
 
 			By("Checking DataVolume exists")
-			err = WaitForDataVolumePhase(*kvClient, namespace.Name, cdiv1.Succeeded, "test-dv")
+			err = framework.WaitForDataVolumePhase(kvClient, namespace.Name, cdiv1.Succeeded, "test-dv")
 			Expect(err).ToNot(HaveOccurred())
 		})
 	})
@@ -138,7 +138,7 @@ var _ = Describe("DV Backup", func() {
 		})
 
 		AfterEach(func() {
-			err := DeleteDataVolume(*kvClient, namespace.Name, dv.Name)
+			err := framework.DeleteDataVolume(kvClient, namespace.Name, dv.Name)
 			Expect(err).ToNot(HaveOccurred())
 
 			err = client.CoreV1().Namespaces().Delete(context.TODO(), sourceNamespace.Name, metav1.DeleteOptions{})
@@ -148,61 +148,61 @@ var _ = Describe("DV Backup", func() {
 
 		})
 
-		It("xxx DataVolume should be restored", func() {
+		It("DataVolume should be restored", func() {
 			var err error
 			By("Creating source DV")
 			sourceVolumeName := "source-volume"
-			srcDvSpec := NewDataVolumeForBlankRawImage(sourceVolumeName, "100Mi", r.StorageClass)
+			srcDvSpec := framework.NewDataVolumeForBlankRawImage(sourceVolumeName, "100Mi", r.StorageClass)
 			srcDvSpec.Annotations[forceBindAnnotation] = "true"
 
 			By(fmt.Sprintf("Creating DataVolume %s", srcDvSpec.Name))
-			srcDv, err := CreateDataVolumeFromDefinition(*kvClient, sourceNamespace.Name, srcDvSpec)
+			srcDv, err := framework.CreateDataVolumeFromDefinition(kvClient, sourceNamespace.Name, srcDvSpec)
 			Expect(err).ToNot(HaveOccurred())
-			err = WaitForDataVolumePhase(*kvClient, srcDv.Namespace, cdiv1.Succeeded, srcDv.Name)
+			err = framework.WaitForDataVolumePhase(kvClient, srcDv.Namespace, cdiv1.Succeeded, srcDv.Name)
 			Expect(err).ToNot(HaveOccurred())
 
 			By("Creating source pod")
-			podSpec := NewPod("source-use-pod", sourceVolumeName, "while true; do echo hello; sleep 2; done")
-			_, err = (*kvClient).CoreV1().Pods(sourceNamespace.Name).Create(context.TODO(), podSpec, metav1.CreateOptions{})
+			podSpec := framework.NewPod("source-use-pod", sourceVolumeName, "while true; do echo hello; sleep 2; done")
+			_, err = (kvClient).CoreV1().Pods(sourceNamespace.Name).Create(context.TODO(), podSpec, metav1.CreateOptions{})
 			Expect(err).ToNot(HaveOccurred())
 			Eventually(func() v1.PodPhase {
-				pod, err := (*kvClient).CoreV1().Pods(sourceNamespace.Name).Get(context.TODO(), podSpec.Name, metav1.GetOptions{})
+				pod, err := (kvClient).CoreV1().Pods(sourceNamespace.Name).Get(context.TODO(), podSpec.Name, metav1.GetOptions{})
 				Expect(err).ToNot(HaveOccurred())
 				return pod.Status.Phase
 			}, 90*time.Second, 2*time.Second).Should(Equal(v1.PodRunning))
 
 			By("Creating clone DV - object under test")
-			dvSpec := NewCloneDataVolume("test-dv", "100Mi", srcDv.Namespace, srcDv.Name, r.StorageClass)
-			dv, err = CreateDataVolumeFromDefinition(*kvClient, namespace.Name, dvSpec)
+			dvSpec := framework.NewCloneDataVolume("test-dv", "100Mi", srcDv.Namespace, srcDv.Name, r.StorageClass)
+			dv, err = framework.CreateDataVolumeFromDefinition(kvClient, namespace.Name, dvSpec)
 
 			By("Creating backup test-backup")
-			err = CreateBackupForNamespace(timeout, backupName, namespace.Name, snapshotLocation, r.BackupNamespace, true)
+			err = framework.CreateBackupForNamespace(timeout, backupName, namespace.Name, snapshotLocation, r.BackupNamespace, true)
 			Expect(err).ToNot(HaveOccurred())
 
-			phase, err := GetBackupPhase(timeout, backupName, r.BackupNamespace)
+			phase, err := framework.GetBackupPhase(timeout, backupName, r.BackupNamespace)
 			Expect(err).ToNot(HaveOccurred())
 			Expect(phase).To(Equal(velerov1api.BackupPhaseCompleted))
 
 			By("Deleting DataVolume")
-			err = DeleteDataVolume(*kvClient, namespace.Name, dv.Name)
+			err = framework.DeleteDataVolume(kvClient, namespace.Name, dv.Name)
 			Expect(err).ToNot(HaveOccurred())
-			ok, err := WaitDataVolumeDeleted(*kvClient, namespace.Name, dv.Name)
+			ok, err := framework.WaitDataVolumeDeleted(kvClient, namespace.Name, dv.Name)
 			Expect(err).ToNot(HaveOccurred())
 			Expect(ok).To(BeTrue())
 
 			By("Deleting source pod")
-			err = (*kvClient).CoreV1().Pods(sourceNamespace.Name).Delete(context.TODO(), podSpec.Name, metav1.DeleteOptions{})
+			err = (kvClient).CoreV1().Pods(sourceNamespace.Name).Delete(context.TODO(), podSpec.Name, metav1.DeleteOptions{})
 			Expect(err).ToNot(HaveOccurred())
 
 			By("Creating restore test-restore")
-			err = CreateRestoreForBackup(timeout, backupName, restoreName, r.BackupNamespace, true)
+			err = framework.CreateRestoreForBackup(timeout, backupName, restoreName, r.BackupNamespace, true)
 			Expect(err).ToNot(HaveOccurred())
 
-			err = WaitForRestorePhase(timeout, restoreName, r.BackupNamespace, velerov1api.RestorePhaseCompleted)
+			err = framework.WaitForRestorePhase(timeout, restoreName, r.BackupNamespace, velerov1api.RestorePhaseCompleted)
 			Expect(err).ToNot(HaveOccurred())
 
 			By("Checking DataVolume exists")
-			err = WaitForDataVolumePhase(*kvClient, namespace.Name, cdiv1.Succeeded, "test-dv")
+			err = framework.WaitForDataVolumePhase(kvClient, namespace.Name, cdiv1.Succeeded, "test-dv")
 			Expect(err).ToNot(HaveOccurred())
 		})
 	})
